et voici les pbuffers \!

+++ src/cairo-dock-dock-manager.c
@@ -386,8 +386,10 @@ gboolean cairo_dock_get_root_dock_position (const gchar *cDockName, CairoDock *p
 	g_key_file_load_from_file (pKeyFile, cConfFilePath, G_KEY_FILE_KEEP_COMMENTS | G_KEY_FILE_KEEP_TRANSLATIONS, &erreur);
 	if (erreur != NULL)
 	{
-		cd_warning ("Attention : %s", erreur->message);
+		cd_warning (erreur->message);
 		g_error_free (erreur);
+		if (! pDock->bIsMainDock)
+			g_free (cConfFilePath);
 		return FALSE;
 	}
 	else
@@ -435,7 +437,7 @@ gboolean cairo_dock_get_root_dock_position (const gchar *cDockName, CairoDock *p
 void cairo_dock_remove_root_dock_config (const gchar *cDockName)
 {
 	gchar *cConfFilePath = g_strdup_printf ("%s/%s.conf", g_cCurrentThemePath, cDockName);
-	if (g_file_test (cConfFilePath, G_FILE_TEST_EXISTS))
+	if (g_file_test (cConfFilePath, G_FILE_TEST_EXISTS) && cDockName && strcmp (cDockName, "cairo-dock.conf") != 0)
 	{
 		g_remove (cConfFilePath);
 	}
